Creates NumberOfEvents.test.js testing NumberOfEvents componentt
A test in NumberOfEvents.test.js to ensure that the NumberOfEvents component contains an element with the role of the textbox.
A test in NumberOfEvents.test.js to ensure that the default value of the input field is 32.
A test in NumberOfEvents.test.js to ensure that the value of the NumberOfEvents component’s textbox has a value that changes accordingly when a user .type()s in it. (Note: If you want to simulate pressing backspace twice and then typing 10, you would write the code await user.type(TextInput_Reference_Variable, '{backspace}{backspace}10');).