imported App component & React Testing Library functions
within(), allows you to use React Testing Library query functions on the passed DOM object. Such query functions are queryByText, queryByRole, and so on—like the query functions you would call on the returned object of render().

As for waitFor(), it’s useful if you need a way to query elements in the page that aren’t rendered immediately. In this case, it will only be possible to render events inside the EventList when the getEvents() function in the App component actually finishes (and thus fetches the list of events objects asynchronously). Notice that expect() is also placed inside waitFor() because it’s tied to what’s returned from this delayed rendering update (i.e., the list of events).