created two new states
EventList component expects a prop named events, and so this is needed for the component to render the actual list of Event components.

currentNOE (current “number of events”) is the app's global state representing the current number of events. Its default value is 32, which is one of the app’s requirements.