initialised local state suggestions
Initialise the local state suggestions to have the default value as the same array as its allLocations prop: using a useEffect hook, because the prop will have an empty array until getEvents() is fully executed in the App component (getEvents is an async piece of code).

In useEffect, use the stringified value of the allLocation prop as a dependency. This way, if there’s a change in it (e.g., an empty array that gets filled), the useEffect code will be re-executed again, ensuring that the local suggestions state is updated.