Added code for when no access token found in localStorage
If no access token is found in localStorage, the app should redirect the user to the Google Authorization screen so they can get their authorization code. Google will then redirect the user back to your app with the code, which you can use to retrieve the access token from the Lambda functions on your authorization server.

-the code checks whether an access token was found.
-if no token is found (!accessToken), the code then checks for an authorization code.
-if no authorization code is found (!code), the user is automatically redirected to the Google Authorization screen, where they can sign in and receive their code.