defined getToken
When token doesn’t exist/is invalid, need to get a new one: redirect users to log in w/ Google so they can be redirected back to our site w/ the code. The new token will be fetched in the code snippet of 'accessToken' function using the function getToken, which is called if the authorisation code is present.