passed setInfoAlert setter function to CitySearch
define when to render the alert. For example, when a user searches for a city and no suggestions are found. In this case, you would need to pass the setInfoAlert setter function to CitySearch to populate it with a string value (the alert message to be displayed) in case the user searches for a city that doesn’t exist.