call/define setInfoAlert
call setInfoAlert inside handleInputChanged function, where user typing is handled and get the list of suggestions.

Once list of filtered locations is returned: checks if list doesn't contain any suggestion. If so,  infoText set as the msg to be displayed in info alert. If list contains suggestions, infoText is set as empty, and alert stays hidden.