implemented offline logic
saves the fetched events data to localStorage for future use:

JSON.stringify(result.events) function call is necessary because events is an array, but localStorage can only store strings.

The JSON.stringify() function will convert the list into a string, allowing it to be stored in localStorage. It will then need to be parsed later using the JSON.parse() function when it’s loaded from localStorage. The (stringified) list is stored under the key lastEvents.